Use deps.edn
The project is split into two domains: client and server. The client gets compiled by shadow into JavaScript, which the server uses to present to the user while running backend services (e.g. querying for social media). Initially, I thought it'd be possible for shadow to handle it (especially since it worked in my editor), but it's become apparent that the client and server need to be run separately.

In deps.edn, a top-level dependency map is listed, while there are separate aliases for Clojure (`:clj`) and ClojureScript (`:cljs`). There is probably some redundancy in this split as shadow analyzes and only includes required dependencies, but this works well.

This split also resolves a dependency problem with Datalevin where it required a dependency it did not have access to (for some reason).