Fix deps
This commit restores `shadow-cljs` as an npm dependency. A prior commit removed it, but it turns out that shadow prints a warning if it's not listed.

In `deps.edn`, `resources` was moved from a top-level path to a path only added when using the `:clj` alias. The `resources/public` folder is used by the server to serve JavaScript files compiled by shadow, which aren't needed by shadow in ClojureScript.