Simplify config
Checks for debug mode have been removed to prevent potential client-server incompatibility. Previously, the `debug?` symbol in the `util` namespace used `goog.DEBUG` in ClojureScript and defaulted to `false` in Clojure. While this "worked", it opened the opportunity to the value of `debug?` differing on runs. For example, under normal runs, the client would be in debug mode (`true`), but the server would always be in non-debug mode (`false`). Although I haven't run into the issue heads-on, it could cause issues in the future, such as if `debug?` were used in the UI under the root component (as hydration may see two different states).

Additionally, the `server.db` namespace has been renamed to `server.database` and config reads Discord keys under a different namespace (`server.remote.discord` instead of just `server`).