Replace Fulcro routing for custom routing
This new routing system uses Reitit and basic query and mutation handling in Fulcro. The server matches valid routes and includes them in the initial state. On the client, the app is hydrated and routing is matched by name with path and query parameters for inputs. The client can route to a different page using the `route` mutation, which uses a remote dedicated to the browser page to update the URL dynamically.

This routing system, in my opinion, is far more simpler than Fulcro's default routing. It has some limitations and is in need of improvement (like creating a `route*` function for the benefit of the server), but so far, it's good.