Reorganize routing
The previous routing system relied heavily on reader conditionals to create separate routers on Clojure and ClojureScript. The client only needs to know the names and potentially a few extra parameters to generate links while the server needs handlers to process them. The previous system worked for some time, but ran into circular dependency errors that were very difficult to deal with. With the previous structure, the two main solutions were to either use reader conditionals in the UI to avoid requiring the server (which would make first page load links not visible, but also likely result in the same issue on the server later on), or to create a base router and extend it on the server. I went with the second solution, though the first one had the perk of links only being clickable once the page was hydrated.

This reorganization, in my opinion, is far simpler than before. There are more problems to solve in the routing system, such as how to load data when navigating, but I'm happy this issue is resolved.