Rework routing
The main feature of this commit is that UI routing is implemented with a union query. The previous implementation relied on pure data (`:route/name`, `:route/params`, etc.) with a `case`, but it often did not work. This implementation has the advantage of not creating a generic format for representing routes (instead reliant on properties), flowing through the data tree much better.