Split base router for UI
The base router (in `route.cljc`) has been split into two routers where routing concerned with the UI is in a separate router. This allows the base router to be independent of UI code, which makes it safe to require from the UI (in `ui.cljc`). So, the dependency graph can be thought of as so:
- `route.cljc` as a base
- `route/ui.cljc` as a router concerned with the UI
- `client/route.cljs` concerned with routing on the client from the UI
- `server/route.clj` concerned with routing on the server for the client from the UI

This split was necessary to resolve a potential circular dependency in the future. As the router was coupled to the UI, it could have been possible in the future for the base to require the UI (via `require`), which would lead to a dependency conflict since the UI requires the UI to generate hrefs. With this change, the UI can safely require the base, be unconcerned about the router dedicated to the UI, and let the other routers consume it. A nice solution.