Add retry capabilities to Discord requests
This commit adds retrying functionality to Discord requests on access tokens. When a token expires (usually in a week), it needs to be refreshed to continue requesting. There was no functionality for this before as the project is still local and in development, but now there is with the `use-token` function. It takes a function that's expected to make an HTTP request to Discord and produce a response. If the response indicates the request was unauthorized (which would mean that the token is invalid), the token is refreshed and the function is called again.