Push Discord and GitHub loading to the server
The client no longer needs to make two separate requests on page load for this. It still needs to for routing (e.g. loading projects), but that's trickier to do, as the server only needs normalizing functions, while the client needs transacting functions (and, in the future, ClojureScript-only functions, such as writing to the local storage). Maybe updating the server router can solve this?